Initialize JRuby's ScriptingContainer lazily
